HeroX_888ad54_exact/requirements.txt
Flask>=2.0
werkzeug
python-telegram-bot==13.15
websockets
requests
python-deriv-api
pandas
numpy
scikit-learn
joblib
beautifulsoup4
lxml
